rename agentic-stack to agentic-api
