+++ tests/test_vllm_cli.py
@@ -1,6 +1,6 @@
 import pytest
 
-from agentic_stack.entrypoints import vllm_cli
+from agentic_api.entrypoints import vllm_cli
 
 
 def _fake_vllm_proc():
@@ -39,7 +39,7 @@ def test_vllm_cli_runs_serve_when_flag_present(monkeypatch: pytest.MonkeyPatch) 
         [
             "serve",
             "model",
-            "--agentic-stack",
+            "--agentic-api",
             "--port",
             "8457",
             "--gateway-port",
@@ -61,7 +61,7 @@ def test_vllm_cli_infers_llm_api_base_from_port(
     monkeypatch.setattr(vllm_cli, "_spawn_vllm", lambda argv: _fake_vllm_proc())
     monkeypatch.setattr(vllm_cli, "run", lambda config: seen.update({"config": config}))
 
-    vllm_cli.main(["serve", "model", "--agentic-stack", "--port", "9999"])
+    vllm_cli.main(["serve", "model", "--agentic-api", "--port", "9999"])
 
     assert seen["config"].llm_api_base == "http://127.0.0.1:9999"
 
@@ -74,7 +74,7 @@ def test_vllm_cli_infers_default_port_when_no_port(
     monkeypatch.setattr(vllm_cli, "_spawn_vllm", lambda argv: _fake_vllm_proc())
     monkeypatch.setattr(vllm_cli, "run", lambda config: seen.update({"config": config}))
 
-    vllm_cli.main(["serve", "model", "--agentic-stack"])
+    vllm_cli.main(["serve", "model", "--agentic-api"])
 
     assert seen["config"].llm_api_base == "http://127.0.0.1:8000"
 
@@ -89,7 +89,7 @@ def test_vllm_cli_normalizes_v1_suffix(monkeypatch: pytest.MonkeyPatch) -> None:
         [
             "serve",
             "model",
-            "--agentic-stack",
+            "--agentic-api",
             "--llm-api-base",
             "http://127.0.0.1:8000/v1",
         ]
@@ -98,11 +98,11 @@ def test_vllm_cli_normalizes_v1_suffix(monkeypatch: pytest.MonkeyPatch) -> None:
     assert seen["config"].llm_api_base == "http://127.0.0.1:8000"
 
 
-def test_vllm_cli_prints_help_for_agentic_stack_help(
+def test_vllm_cli_prints_help_for_agentic_api_help(
     capsys: pytest.CaptureFixture[str],
 ) -> None:
     with pytest.raises(SystemExit) as excinfo:
-        vllm_cli.main(["serve", "model", "--agentic-stack", "--help"])
+        vllm_cli.main(["serve", "model", "--agentic-api", "--help"])
 
     assert excinfo.value.code == 0
     assert "--gateway-port" in capsys.readouterr().out
